Retaining matrix structure
When you drop the exponentiated column from a one-row matrix (ie, a Cox model with only 1 continuous or binary independent variable), the object returned is a numeric vector, and `fix_data_frame` can't figure out what to do with it.  Adding the `drop=FALSE` argument keeps the matrix in tact.