bootstrapped temporal ERGM tidier
This really is a wrap-up of the `confint` method of the `xergm`
package. There is nothing very fancy about `term` or `mtergm` models,
and no easy-to-obtain GOF statistics, so the only method here is
`tidy`, which optionally exponentiates the coefficients and confidence
intervals.

I just noticed something: `quick=TRUE` does not take `exponentiate`
into account, even though both arguments can be passed together.