Add argument to `tidy.SpatialPolygonsDataFrame` to retain variables from `SpatialPolygonsDataFrame`
`SpatialPolygonsDataFrame` combine polygon information with a metadata in a `data.frame`. In many situations, it's useful to keep some of these variables. For example, it makes plotting maps much easier. My changes use the `...` argument with the same syntax as `dplyr::select` to specify the variables the user wants to keep. Obviously, a tidy SpatialPolygonsDataFrame is not a very efficient way to store this data. That is why it makes more sense that the user explicitly selects the variables she/he wants to keep. `...` is just passed to `dplyr::select` so all the helper functions work as well (e.g. `starts_with`, `ends_with` etc).