Improve the tidy method for "ts" time series objects.
This current implementation merely calls out to as.data.frame(). This is
problematic for two reasons: (i) the result is not "tidy" for multivariate ts
objects, in the sense of having one row per observation, and (ii) the result has
meaningful row names.

This commit provides a new implementation, based on the approach used in
`zoo_tidiers`. The tidy method now returns a data frame with `index`, `series`,
and `value` columns.

The new documentation contains univariate and multivariate "ts" examples.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>