Make tidy() and glance() handle survfit() objects with missing CIs
Running tidy() on a ‘survfit’ or ‘survfitms’ object which
didn’t have a confidence interval (i.e. was created with
conf.type="none" or conf.int=FALSE) resulted in an error message.
This commit fixes this bug (fixes #305). Now the confidence
interval is only included in the tidy data frame if the
‘survfit’ object actually includes a confidence interval.
Also, the order of the confidence limit in the resulting
object is changed, from ‘conf.high, conf.low’ to the more
natural ‘conf.low, conf.high’, for consistency with other
broom functions (e.g. tidy.lm and glance.survfit).

A similar bug was present when running glance() on a ‘survfit’
object which didn’t have a confidence interval. It resulted
in the median and standard error of the restricted mean being
mislabelled as the upper and lower confidence limits. This
commit also fixes this bug (fixed #304).

The documentation for the two functions is also updated
to reflect the changes.