multivariate support for tidy.acf
`acf` and `pacf` support multivariate inputs, in which case the current
code winds up providing a data.frame with lots of poorly differentiated
lag and acf columns. This commit uses snames to keep acf and lag to
single columns with two additional columns describing the time series
in 'row' and 'col' positions as they would be displayed in the base
acf plot.