Changed from parsing strata string to finding name matches.
The first attempt at implementation split the "strata" string at commas, and used equals signs as delineators for new values.  Those broke if you included a comma or equals sign in the variable name (like interaction(x, y) or sex == "male").  This new approach does 3 things: (1) it gets the names of the terms in the equation with "terms", (2) it finds the locations of those values in the "strata" string, and (3) it extracts the values in between those terms as the values of the right hand side (rhs).