Update strata parser.
Strata strings are pasted together in a fixed width format, "name1=value1, name2=value2, name3=value3".  The previous version of the parser took advantage of the fixed width format  (= and , are at the same positions).  The new version uses regular expressions that look for "=" and ",", and fails with a warning if there are too many "=" or ",".

Note, too, that the previous version used x$call, but we cannot rely on x$call having a formula with the strata values.