Update survival-survfit-tidiers.R
I found that the values of standard error of a sufvfit() object (Kaplan-Meier estimate) depends on whether is is extracting from the sufvit object directly or extracting from the summary(survfit) object. I think the standard error estimates of summary(survfit) object are correct, but tidy(survfit) uses standard error of survfit object. So I suggest to make the changes so tidy() provides correct standard error of Kaplan-Meier estimates.