Revert rma confidence intervals to .conf.low/.conf.high; change prediction to .pred.low/.pred.high
The columns previously labeled "cr.lb" and "cr.ub" in rma objects are now called "pi.lb" and "pi.ub". The previous label was misleading and inaccurate--these values are not "credibility intervals" as used in the meta-analysis literature or "credible intervals" as used in Bayesian analysis. They are prediction intervals ala lm(). The change in metafor was to avoid some common confusion and misinterpretation among metafor users; this change to "pred" instead of "cred" follows suit.