"soft-support" lfe::felm objects
lfe was archived from CRAN recently, but we suspect that it might make it back to CRAN at some point soon.

this commit supports lfe objects without referencing the package at all / taking it on as a Suggests. bumps all references to the lfe package itself, (temporarily) deprecating tests and examples and removing the package from the description, but continuing to export the tidier methods.