Check that outputs of xyz lists are numeric
x,y,z lists are usually suitable for persp() and image(); for those, all elements must be numeric so it makes sense to check that they are numeric here.

check_type() is not appropriate here since we just want to check for the numeric nature, not necessarily distingish between integer and double for example.