Improve performance of tidy.lm and tidy.glm for full-rank fits
tidy.lm and tidy.glm use summary(x)$coefficients, then join with coef(x)
to handle the case where the fit is rank deficient and some terms with
NA coefficients are missing from summary(x)$coefficients. However, in a
simple experiment, the left_join() consumes a disproportionate amount of
time. Only run the left_join() when the fit is rank-deficient and it is
necessary, and add tests to ensure the rows are added correctly for
rank-deficient fits.

An example:

```
library(broom)
library(profvis)

profvis(replicate(1000, tidy(lm(dist ~ speed, data = cars))))
```

Before this change, the expression ran in 3470 ms, of which 2570
ms (74%) was in tidy.lm() and 1590 ms (46%) was in
left_join.data.frame(). After this change, the expression ran in 1740
ms, of which 1010 ms was in tidy.lm(), cutting 60% from the total time
in tidy.lm(). This is useful for simulation studies that use tidy() to
get estimates for each simulated fit.